requirements/dev.txt
fastapi
bcrypt
cryptography
psycopg2-binary
PyMySQL
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pytest
pydantic
httpx
pytest-asyncio
aiofiles
jinja2
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
alembic